backend-flask/requirements.txt
flask==3.0.2
flask-cors==5.0.0
pillow==10.3.0
numpy==1.26.4
pytest==8.3.3
torch>=2.6.0
torchvision>=0.21.0
transformers==4.44.2
tokenizers==0.22.0
scikit-image==0.24.0
opencv-python-headless==4.10.0.84
scipy==1.13.1
PyWavelets>=1.4.1
